# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Simple API for to-do-list management posts on social media",
        "title": "Sosmed Todolist API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/platforms": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Platform"
                ],
                "summary": "Get all platforms",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Keyword to search",
                        "name": "keyword",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of entities per page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully fetched all platforms",
                        "schema": {
                            "$ref": "#/definitions/platforms.ListofPlatforms"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/httpres.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/httpres.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Platform"
                ],
                "summary": "Create a new platform",
                "parameters": [
                    {
                        "description": "Platform details",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/platforms.PlatformRequestPayload"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Platform successfully created",
                        "schema": {
                            "$ref": "#/definitions/httpres.BaseResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/httpres.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/httpres.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/platforms/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Platform"
                ],
                "summary": "Get a single platform by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Platform ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully fetched the platform",
                        "schema": {
                            "$ref": "#/definitions/platforms.PlatformDetails"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/httpres.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Platform not found",
                        "schema": {
                            "$ref": "#/definitions/httpres.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/httpres.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Platform"
                ],
                "summary": "Update an existing platform",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Platform ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated platform details",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/platforms.PlatformRequestPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Platform updated successfully",
                        "schema": {
                            "$ref": "#/definitions/httpres.BaseResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/httpres.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Platform not found",
                        "schema": {
                            "$ref": "#/definitions/httpres.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/httpres.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Platform"
                ],
                "summary": "Delete a platform by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Platform ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Platform deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/httpres.BaseResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/httpres.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Platform not found",
                        "schema": {
                            "$ref": "#/definitions/httpres.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/httpres.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "httpres.BaseResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "httpres.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "httpres.ListPagination": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer",
                    "example": 100
                },
                "page": {
                    "type": "integer",
                    "example": 1
                },
                "total_page": {
                    "type": "integer",
                    "example": 10
                }
            }
        },
        "platforms.ListofPlatforms": {
            "type": "object",
            "properties": {
                "meta": {
                    "$ref": "#/definitions/httpres.ListPagination"
                },
                "platforms": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/platforms.PlatformDetails"
                    }
                }
            }
        },
        "platforms.PlatformDetails": {
            "type": "object",
            "properties": {
                "platform": {
                    "type": "string"
                },
                "platform_id": {
                    "type": "integer"
                }
            }
        },
        "platforms.PlatformRequestPayload": {
            "type": "object",
            "required": [
                "platform"
            ],
            "properties": {
                "platform": {
                    "type": "string",
                    "maxLength": 200,
                    "minLength": 1
                }
            }
        },
...
    }
}